feat(docgen): docgen으로 디렉토리 명 및 구조 변경
