+++ packages/core/src/docgen/render/contexts/curl-request.spec.ts
@@ -1,4 +1,4 @@
-import type { DocumentSnapshot } from "@/builders";
+import type { DocumentSnapshot } from "@/docgen/builders";
 import { MissingFieldError } from "@/errors";
 import { describe, expect, it } from "vitest";
 import { buildCurlContext } from "./curl-request";
