+++ packages/core/src/docgen/render/contexts/curl-request.ts
@@ -1,4 +1,4 @@
-import type { DocumentSnapshot } from "@/builders";
+import type { DocumentSnapshot } from "@/docgen/builders";
 import { MissingFieldError } from "@/errors";
 import type { HttpBody, HttpCookies, HttpHeaders, HttpMethod, HttpQuery } from "@/types";
 import { compact, isEmpty, join, toPairs } from "es-toolkit/compat";
