+++ packages/core/src/docgen/render/contexts/request-body.spec.ts
@@ -1,4 +1,4 @@
-import type { DocumentSnapshot } from "@/builders";
+import type { DocumentSnapshot } from "@/docgen/builders";
 import { describe, expect, it } from "vitest";
 import { buildRequestBodyContext } from "./request-body";
