+++ packages/core/src/docgen/render/document/index.ts
