+++ packages/core/src/docgen/render/document/openapi-renderer.ts
