+++ packages/core/src/docgen/render/snippet/curl-request.ts
@@ -1,18 +1,18 @@
-import type { DocumentSnapshot } from "@/builders";
-import { buildRequestFieldsContext } from "../contexts";
+import type { DocumentSnapshot } from "@/docgen/builders";
+import { buildCurlContext } from "../contexts";
 import type { TemplateStore } from "../template";
 import { createTemplateRenderer } from "../template/template-factory";
 import type { SnippetRenderer } from "./snippet-renderer.type";
 
-export class RequestFieldsSnippetRenderer implements SnippetRenderer {
-    readonly templateName = "request-fields";
+export class CurlSnippetRenderer implements SnippetRenderer {
+    readonly templateName = "curl-request";
 
     constructor(private readonly store: TemplateStore) {}
 
     async render(snapshot: DocumentSnapshot): Promise<string> {
         const { extension, content } = this.store.get(this.templateName);
         const renderer = createTemplateRenderer(extension);
-        const context = buildRequestFieldsContext(snapshot);
+        const context = buildCurlContext(snapshot);
 
         const result = await renderer.render(content, context);
         return result;
