+++ packages/core/src/docgen/render/snippet/http-request.ts
@@ -1,18 +1,18 @@
-import type { DocumentSnapshot } from "@/builders";
-import { buildRequestCookiesContext } from "../contexts";
+import type { DocumentSnapshot } from "@/docgen/builders";
+import { buildHttpRequestContext } from "../contexts";
 import type { TemplateStore } from "../template";
 import { createTemplateRenderer } from "../template/template-factory";
 import type { SnippetRenderer } from "./snippet-renderer.type";
 
-export class RequestCookiesSnippetRenderer implements SnippetRenderer {
-    readonly templateName = "request-cookies";
+export class HttpRequestSnippetRenderer implements SnippetRenderer {
+    readonly templateName = "http-request";
 
     constructor(private readonly store: TemplateStore) {}
 
     async render(snapshot: DocumentSnapshot): Promise<string> {
         const { extension, content } = this.store.get(this.templateName);
         const renderer = createTemplateRenderer(extension);
-        const context = buildRequestCookiesContext(snapshot);
+        const context = buildHttpRequestContext(snapshot);
 
         const result = await renderer.render(content, context);
         return result;
