+++ packages/core/src/docgen/render/snippet/request-headers.ts
@@ -1,18 +1,18 @@
-import type { DocumentSnapshot } from "@/builders";
-import { buildCurlContext } from "../contexts";
+import type { DocumentSnapshot } from "@/docgen/builders";
+import { buildRequestHeadersContext } from "../contexts";
 import type { TemplateStore } from "../template";
 import { createTemplateRenderer } from "../template/template-factory";
 import type { SnippetRenderer } from "./snippet-renderer.type";
 
-export class CurlSnippetRenderer implements SnippetRenderer {
-    readonly templateName = "curl-request";
+export class RequestHeadersSnippetRenderer implements SnippetRenderer {
+    readonly templateName = "request-headers";
 
     constructor(private readonly store: TemplateStore) {}
 
     async render(snapshot: DocumentSnapshot): Promise<string> {
         const { extension, content } = this.store.get(this.templateName);
         const renderer = createTemplateRenderer(extension);
-        const context = buildCurlContext(snapshot);
+        const context = buildRequestHeadersContext(snapshot);
 
         const result = await renderer.render(content, context);
         return result;
