+++ packages/core/src/docgen/render/snippet/snippet-renderer.type.ts
@@ -1,4 +1,4 @@
-import type { DocumentSnapshot } from "@/builders";
+import type { DocumentSnapshot } from "@/docgen/builders";
 
 export interface SnippetRenderer {
     readonly templateName: string;
