+++ packages/core/src/docgen/render/template/index.ts
