+++ packages/core/src/docgen/render/template/template-factory.ts
