+++ packages/core/src/docgen/render/template/template-renderer.type.ts
