+++ packages/core/src/docgen/render/template/template-store.ts
