+++ packages/core/src/docgen/writers/doc-writer.interface.ts
