+++ packages/core/src/docgen/writers/index.ts
@@ -1,0 +1,2 @@
+export * from "./doc-writer.interface";
+export * from "./local-doc-writer";
