+++ packages/core/src/index.ts
@@ -1,8 +1,8 @@
-export { docRequest } from "./builders/document-builder";
+export { docRequest } from "./docgen/builders/document-builder";
 export { setNRestDocsConfig, getNRestDocsConfig } from "./config/config";
 
 export * from "./errors";
-export * from "./builders";
+export * from "./docgen/builders";
 export * from "./descriptors";
 export * from "./inputs";
 export * from "./types";
